ducktype method caller implementation
